requirements.txt
fastapi==0.108.0
Jinja2==3.1.2
sqlmodel==0.0.14
python-dotenv==1.0.0
uvicorn==0.25.0
pydantic==2.5.3
bcrypt==4.1.2
email-validator==2.1.0.post1
pytz==2023.3.post1
python-multipart==0.0.6
python-jose==3.3.0
cryptography==41.0.7
python-magic==0.4.27
